php works in rs

+++ php/src/main.rs
@@ -6,7 +6,7 @@ use std::env;
 fn main() {
     let args: Vec<_> = env::args().collect();
     assert!(args.len() == 2, "Usage: cargo run n");
-    let n: u64 = args[1].parse().unwrap();
+    let n: u32 = args[1].parse().unwrap();
     println!("{}", n);
     println!("---");
     println!("Solving PHP for {} pigeons and {} holes", n + 1, n);
